#184 Fix findActivePopupMenu to work with cascading menus.
findActivePopupMenu does not process cascading menus or popup menus. The
usual manner in creating menu is adding JMenuItem and JMenu
(popup/cascade object in a JPopupMenu) to create menus. Since day one of
FEST it appears this has not been clear in BasicRobot and in the test
class BasicRobot_findActivePopupMenu.

This fixes the code more fully by knowing about JPopupMenu, JMenuItem
and JMenu interaction.